Add `web/templates/page/test.html.eex`
We have a route. We created a new controller action. We have made modifications to the main application view. Now all we need is a new template to display the string we get from `handler_info/1`. Let's create a new one at `web/templates/page/test.html.eex`.

    # web/templates/page/test.html.eex
    <div class="jumbotron">
      <p><%= handler_info @conn %></p>
    </div>

Notice that `@conn` is available to us in the template for free via the `assigns` map.

If we visit [localhost:4000/test](http://localhost:4000/test), we will see that our page is brought to us by `Elixir.HelloPhoenix.PageController.test`.

We can define functions in any individual view in `web/views`. Functions defined in an individual view will only be available to templates which that view renders. For example, functions like our `handler_info` above, will only be available to templates in `web/templates/page`.