Add a list to web/templates/page/test.html.eex
##### Displaying Lists

So far, we've only displayed singular values in our templates - strings here, and integers in other guides. How would we approach displaying all the elements of a list?

The answer is that we can use Elixir's list comprehensions.

Now that we have a function, visible to our template, that returns a list of keys in the `conn` struct, all we need to do is modify our `web/templates/page/test.html.eex` template a bit to display them.

We can add a header and a list comprehension like this.

    # web/templates/page/test.html.eex
    <div class="jumbotron">
      <p><%= handler_info @conn %></p>

      <h3>Keys for the conn Struct</h3>

      <%= for key <- connection_keys @conn do %>
        <p><%= key %></p>
      <% end %>
    </div>

We use the list of keys returned by the `connection_keys` function as the source list to iterate over. Note that we need the `=` in both `<%=` - one for the top line of the list comprehension and the other to display the key. Without them, nothing would actually be displayed.

When we visit [localhost:4000/test](http://localhost:4000/test) again, we see all the keys displayed.