Add web/templates/page/key.html.eex
##### Render templates within templates

In our list comprehension example above, the part that actually displays the values is quite simple.

```html+eex
<p><%= key %></p>
```
We are probably fine with leaving this in place. Quite often, however, this display code is somewhat more complex, and putting it in the middle of a list comprehension makes our templates harder to read.

The simple solution is to use another template! Templates are just function calls, so like regular code, composing your greater template by small, purpose-built functions can lead to clearer design. This is simply a continuation of the rendering chain we have already seen. Layouts are templates into which regular templates are rendered. Regular templates may have other templates rendered into them.

Let's turn this display snippet into its own template. Let's create a new template file at `web/templates/page/key.html.eex`, like this.

    # web/templates/page/key.html.eex
    <p><%= @key %></p>

We need to change `key` to `@key` here because this is a new template, not part of a list comprehension. The way we pass data into a template is by the `assigns` map, and the way we get the values out of the `assigns` map is by referencing the keys with a preceding `@`.