Explicitly use `HelloPhoenix.PageView` in page/test.html.eex
##### Shared Templates Across Views

Often, we find that small pieces of data need to be rendered the same way in different parts of the application. It's a good practice to move these templates into their own shared directory to indicate that they ought to be available anywhere in the app.

Let's move our template into a shared view.

`key.html.eex` is currently rendered by the `HelloPhoenix.PageView` module, but we use a render call which assumes that the current view model is what we want to render with. We could make that explicit, and re-write it like this:

    # web/templates/page/test.html.eex
    <div class="jumbotron">
      ...

      <%= for key <- connection_keys @conn do %>
        <%= render HelloPhoenix.PageView, "key.html", key: key %>
      <% end %>
    </div>